API: fix RX gain having wrong value in SetGain() function

+++ src/API/lms7_api.cpp
@@ -451,7 +451,7 @@ API_EXPORT int CALL_CONV LMS_SetGaindB(lms_device_t *device, bool dir_tx,
                                                 size_t chan,unsigned gain)
 {
     lime::LMS7_Device* lms = CheckDevice(device, chan);
-    return lms ? lms->SetGain(dir_tx,chan,dir_tx ? gain : int(gain-12)) : -1;
+    return lms ? lms->SetGain(dir_tx,chan,dir_tx ? gain : double(gain-12.0f)) : -1;
 }
 
 API_EXPORT int CALL_CONV LMS_GetNormalizedGain(lms_device_t *device, bool dir_tx, size_t chan,float_type *gain)
